build: convert rollup configs to .cjs with dynamic imports
- Rename rollup.config.js to rollup.config.cjs
- Rename rollup.config.serverless.js to rollup.config.serverless.cjs
- Convert require() statements to dynamic import() syntax
- Use async IIFE pattern for CommonJS compatibility
- Add rollup-plugin-esbuild for consistent TypeScript compilation
- Remove deprecated tsconfig.serverless.json
- Update package.json build scripts to reference .cjs files
- Fix serverless tests to use valid-domain.org instead of example.com
- Add biome lint rule exception for callback returns
- Update dependencies: esbuild, rollup-plugin-esbuild, tslib
