+++ package.json
@@ -67,10 +67,9 @@
     "docs/SERVERLESS.md"
   ],
   "scripts": {
-    "build": "rm -rf dist && rollup -c rollup.config.js && yarn build:serverless",
-    "build:main": "rollup -c rollup.config.js",
-    "build:serverless": "rollup -c rollup.config.serverless.js",
-    "build:tsc": "tsc -p . && tsc -p tsconfig.serverless.json",
+    "build": "rm -rf dist && rollup -c rollup.config.cjs && rollup -c rollup.config.serverless.cjs",
+    "build:main": "rollup -c rollup.config.cjs",
+    "build:serverless": "rollup -c rollup.config.serverless.cjs",
     "lint": "biome check",
     "lint:fix": "biome check --write",
     "format": "biome format --write",
@@ -100,10 +99,12 @@
     "@types/jest": "^30.0.0",
     "@types/node": "24.3.0",
     "@types/sinon": "17.0.4",
+    "esbuild": "^0.25.9",
     "husky": "9.1.7",
     "jest": "^30.1.1",
     "lint-staged": "16.1.5",
     "rollup": "^4.49.0",
+    "rollup-plugin-esbuild": "^6.2.1",
     "sinon": "21.0.0",
     "ts-jest": "^29.4.1",
     "tslib": "^2.8.1",
