--- rollup.config.js
@@ -1,51 +1,0 @@
-import typescript from '@rollup/plugin-typescript';
-import json from '@rollup/plugin-json';
-import resolve from '@rollup/plugin-node-resolve';
-import commonjs from '@rollup/plugin-commonjs';
-import { readFileSync } from 'fs';
-
-const pkg = JSON.parse(readFileSync('./package.json', 'utf8'));
-
-export default {
-  input: 'src/index.ts',
-  output: [
-    {
-      file: pkg.main,
-      format: 'cjs',
-      sourcemap: true,
-    },
-    {
-      file: pkg.module || 'dist/index.esm.js',
-      format: 'es',
-      sourcemap: true,
-    },
-  ],
-  external: [
-    ...Object.keys(pkg.dependencies || {}),
-    ...Object.keys(pkg.peerDependencies || {}),
-    'node:dns',
-    'node:net',
-    'node:tls',
-    'node:crypto',
-    'node:fs',
-    'node:path',
-    'node:util',
-    'node:stream',
-  ],
-  plugins: [
-    json({
-      compact: true,
-      preferConst: true,
-    }),
-    resolve({
-      preferBuiltins: true,
-    }),
-    commonjs(),
-    typescript({
-      tsconfig: './tsconfig.json',
-      declaration: true,
-      declarationDir: './dist',
-      exclude: ['**/*.test.ts', '**/*.spec.ts', '__tests__/**/*'],
-    }),
-  ],
-};
