--- tsconfig.serverless.json
@@ -1,26 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist/serverless",
-    "module": "ESNext",
-    "target": "ES2020",
-    "lib": ["ES2020", "DOM"],
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "esModuleInterop": true,
-    "declaration": true,
-    "sourceMap": true,
-    "removeComments": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "types": []
-  },
-  "include": [
-    "src/serverless/**/*",
-    "src/types.ts",
-    "src/disposable-email-providers.json",
-    "src/free-email-providers.json"
-  ],
-  "exclude": ["node_modules", "dist", "__tests__"]
-}
